fix passive data sockets so they work in many basic setups
We were always asking the client to connect to "127.0.0.1:port", which
works great in testing on a local machine, but isn't much use in the
real world.

Now we open the passive data socket on the same server IP that the
control socket is open on. That will work in a standard server setup,
but isn't enough to work with a load balanced (or container) setup.
We'll fix that next.