update tests to work with current master
This PR was opened 5 years ago, and master has moved on a bit since then.

Unfortunately I had to change the signature of one public function:
NewDirItem(). We were always setting the modTime to time.Now(), but
that's silly. It's hard to test, and we should leave it up to the driver
implementation to decide if they need a real time.